이수정 feat(Weapon): 맵 밖, 드럼통체크 /fix(Weapon): 거리 비례 데미지 비율 고정, 조준선 degree 변경, F3탄퍼짐 오류 해결/ feat(Render): -180 degree angle 랜더오류 해결

+++ GameEngineContents/WeaponSheep.cpp
@@ -344,7 +344,7 @@ void WeaponSheep::SheepJump(float _DeltaTime)
 
 void WeaponSheep::Timer()
 {
-    CurTime = clock();
+    CurTime = static_cast<float>(clock());
 
     TimeCount = (CurTime - PrevTime) / 1000.0f;
